Update tools_remote deps for bazel 6.2 (#30)
Bazel 6.2 has introduced warnings promoted to errors via jdk flags that
prevent grpc 1.44.1 from working. Upgrading that to 1.55.1

This introduced an incompatibility with rules_skylib. Upgrading that to
1.4.1 from 0.9.0.
Also incompatible was rules_jvm_external, which was upgraded to 5.2 from
4.2.
Lastly, absl as used by protobuf at the new grpc version is incompatible
with < c++14. Introduced a .bazelrc as a poor excuse for toolchain
injection. Bazel will default to c++14 (from c++0x) with a release that
includes https://github.com/bazelbuild/bazel/pull/18280. An issue to
remove this file will follow.

Tagged the build with .bazelversion to avoid accidental consumption of
incompatible bazels with bazelisk.

Fixes #28